Move files to root and fix date test to use time.Time.
