+++ amf0_decode.go
