+++ amf0_encode.go
