+++ amf3_encode.go
