+++ amf_test.go
@@ -4,6 +4,7 @@ import (
 	"bytes"
 	"reflect"
 	"testing"
+	"time"
 )
 
 func TestEncodeAMF0(t *testing.T) {
@@ -37,7 +38,7 @@ func TestEncodeAMF0(t *testing.T) {
 			0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0x00, 0x01, 0x32, 0x00, 0x7f, 0xef, 0xff, 0xff, 0xff, 0xff,
 			0xff, 0xff, 0x02, 0x00, 0x01, 0x33, 0x02, 0x00, 0x05, 0x74, 0x68, 0x72, 0x65, 0x65, 0x02, 0x00,
 			0x01, 0x34, 0x05, 0x02, 0x00, 0x01, 0x35, 0x01, 0x01}},*/
-		{Amf0Date(353464561), []byte{0x0b, 0x41, 0xb5, 0x11, 0x70, 0xf1, 0x00, 0x00, 0x00, 0x00, 0x00}},
+		{time.Unix(123456789, 123456789), []byte{0x0b, 0x00, 0x00, 0x00, 0x1c, 0xbe, 0x99, 0x1a, 0x83, 0x00, 0x00}},
 		{[]interface{}{"one", "two", "three"}, []byte{0x0a, 0x00, 0x00, 0x00, 0x03, 0x00, 0x00, 0x00,
 			0x00, 0x02, 0x00, 0x03, 0x6f, 0x6e, 0x65, 0x02, 0x00, 0x03, 0x74, 0x77, 0x6f, 0x02, 0x00,
 			0x05, 0x74, 0x68, 0x72, 0x65, 0x65}},
@@ -52,7 +53,7 @@ func TestEncodeAMF0(t *testing.T) {
 	for _, c := range cases {
 		got := EncodeAMF0(c.in)
 		if bytes.Compare(got, c.want) != 0 {
-			t.Errorf("EncodeAMF0(%q) == %q, want %q", c.in, got, c.want)
+			t.Errorf("EncodeAMF0(%#v) == %#v, want %#v", c.in, got, c.want)
 		}
 	}
 }
@@ -88,7 +89,7 @@ func TestDecodeAMF0(t *testing.T) {
 			0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0x00, 0x01, 0x32, 0x00, 0x7f, 0xef, 0xff, 0xff, 0xff, 0xff,
 			0xff, 0xff, 0x02, 0x00, 0x01, 0x33, 0x02, 0x00, 0x05, 0x74, 0x68, 0x72, 0x65, 0x65, 0x02, 0x00,
 			0x01, 0x34, 0x05, 0x02, 0x00, 0x01, 0x35, 0x01, 0x01}},*/
-		{[]byte{0x0b, 0x41, 0xb5, 0x11, 0x70, 0xf1, 0x00, 0x00, 0x00, 0x00, 0x00}, Amf0Date(353464561)},
+		{[]byte{0x0b, 0x00, 0x00, 0x00, 0x1c, 0xbe, 0x99, 0x1a, 0x83, 0x00, 0x00}, time.Unix(123456789, 123000000)},
 		{[]byte{0x0a, 0x00, 0x00, 0x00, 0x03, 0x00, 0x00, 0x00,
 			0x00, 0x02, 0x00, 0x03, 0x6f, 0x6e, 0x65, 0x02, 0x00, 0x03, 0x74, 0x77, 0x6f, 0x02, 0x00,
 			0x05, 0x74, 0x68, 0x72, 0x65, 0x65}, []interface{}{"one", "two", "three"}},
